+++ include/synth/synthbase.h
@@ -42,6 +42,7 @@ public:
 	virtual void HandleMIDISysExMessage(const u8* pData, size_t nSize) = 0;
 	virtual bool IsActive() = 0;
 	virtual void AllSoundOff() = 0;
+	virtual void SetMasterVolume(u8 nVolume) = 0;
 	virtual size_t Render(s16* pOutBuffer, size_t nFrames) = 0;
 	virtual size_t Render(float* pOutBuffer, size_t nFrames) = 0;
 	virtual u8 GetChannelVelocities(u8* pOutVelocities, size_t nMaxChannels) = 0;